Correção da classe MDFInfoModalRodoviarioInfContrato que petence à classe MDFInfoModalRodoviarioInfContratante e inclusão de transformer de dois enums de MDF-e
